Combine fewsnet processing and few improvements
